♻️ Refactor: 모달 템플릿 컴포넌트 분리 및 JSDOC 수정
