+++ src/components/dialog/ButtonArea.tsx
@@ -1,0 +1,6 @@
+import type { DialogProp } from './types';
+
+const ButtonArea = ({ children, className }: DialogProp) => {
+  return <div className={className}>{children}</div>;
+};
+export default ButtonArea;
