+++ src/components/dialog/Close.tsx
@@ -1,0 +1,7 @@
+import type { DialogCloseProp } from './types';
+
+const Close = ({ setToggleModal }: DialogCloseProp) => {
+  return <button onClick={() => setToggleModal()}>x</button>;
+};
+
+export default Close;
