+++ src/components/dialog/Content.tsx
@@ -1,0 +1,7 @@
+import type { DialogProp } from './types';
+
+const Content = ({ children, className }: DialogProp) => {
+  return <div className={className}>{children}</div>;
+};
+
+export default Content;
